NO-TICKET: execution-engine: rename gs -> global_state
