+++ execution-engine/storage/src/global_state/mod.rs
